% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{lator_endpoint_library}
\alias{lator_endpoint_library}
\title{List built-in templates and registered endpoint definitions}
\usage{
lator_endpoint_library(workspace = NULL)
}
\arguments{
\item{workspace}{Optional unlocked workspace.}
}
\description{
List built-in templates and registered endpoint definitions
}
