% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{lator_endpoint_validate}
\alias{lator_endpoint_validate}
\title{Validate an endpoint definition}
\usage{
lator_endpoint_validate(endpoint)
}
\arguments{
\item{endpoint}{Endpoint object.}
}
\value{
The validated endpoint.
}
\description{
Validate an endpoint definition
}
