% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.R
\name{lator_patient_add_events}
\alias{lator_patient_add_events}
\title{Add several event records}
\usage{
lator_patient_add_events(patient, events)
}
\arguments{
\item{patient}{Patient record.}

\item{events}{Data frame or list of named event argument lists.}
}
\value{
Updated patient record.
}
\description{
Add several event records
}
