% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.R
\name{lator_patient_endpoint_get}
\alias{lator_patient_endpoint_get}
\title{Retrieve a medication-specific endpoint preference}
\usage{
lator_patient_endpoint_get(patient, drug)
}
\arguments{
\item{patient}{Patient record.}

\item{drug}{Medication name.}
}
\value{
A therapy-profile list, or `NULL`.
}
\description{
Retrieve a medication-specific endpoint preference
}
