% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.R
\name{lator_patient_get}
\alias{lator_patient_get}
\title{Load a patient}
\usage{
lator_patient_get(workspace, patient_id)
}
\arguments{
\item{workspace}{Unlocked workspace.}

\item{patient_id}{Pseudonym.}
}
\description{
Load a patient
}
