% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.R
\name{lator_patient_list}
\alias{lator_patient_list}
\title{List pseudonymous patient records}
\usage{
lator_patient_list(workspace)
}
\arguments{
\item{workspace}{Unlocked workspace.}
}
\value{
Patient catalogue data frame.
}
\description{
List pseudonymous patient records
}
