% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui-widget.R
\name{liberator_workbench}
\alias{liberator_workbench}
\title{LibeRator React workbench widget}
\usage{
liberator_workbench(
  payload,
  input_id = "liberator_workbench",
  width = NULL,
  height = "100vh",
  elementId = NULL
)
}
\arguments{
\item{payload}{Workbench payload generated by the Shiny application.}

\item{input_id}{Shiny event prefix.}

\item{width, height}{Widget dimensions.}

\item{elementId}{Optional widget element id.}
}
\description{
LibeRator React workbench widget
}
