% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui-widget.R
\name{renderLiberatorWorkbench}
\alias{renderLiberatorWorkbench}
\title{Render a LibeRator workbench}
\usage{
renderLiberatorWorkbench(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Widget expression.}

\item{env}{Evaluation environment.}

\item{quoted}{Whether expression is quoted.}
}
\description{
Render a LibeRator workbench
}
