% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{lator_endpoint_register}
\alias{lator_endpoint_register}
\title{Register an encrypted endpoint definition}
\usage{
lator_endpoint_register(workspace, endpoint, actor = "local-session")
}
\arguments{
\item{workspace}{Unlocked workspace.}

\item{endpoint}{Endpoint to register.}

\item{actor}{Audit actor.}
}
\description{
Register an encrypted endpoint definition
}
